+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to Dexter will be documented in this file.
 
+## [UNRELEASED]
+- TeddySwap integration
+
 ## [v4.2.0]
 - Fix WR price impact formula for 0 decimals
 - Rename Asset identifier function